requirements.txt
flask>2,<3
Flask-Cors==4.0.0
Werkzeug==2.2.2
openpyxl==3.1.2
google-cloud-storage==2.7.0
google-cloud-workflows==1.10.0
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
pandas==2.0.3
SQLAlchemy==2.0.29
